backend/face_recognition/requirements.txt
# Face Recognition Backend Requirements
# Install these packages to run the face recognition API server

flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
Pillow==10.1.0
numpy==1.24.3
